ai_automation_app/backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.4.2
boto3==1.34.0
python-dotenv==1.0.0
mangum==0.17.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
aws-lambda-powertools==2.30.2
structlog==23.2.0
openai==1.6.0
pydub==0.25.1
aiobotocore==2.7.0
requests==2.31.0